Exporting arrays almost complete

+++ Source/ATA_UU/SavedStatesQ.h
@@ -22,7 +22,7 @@ struct FStateAndActions
 	GENERATED_USTRUCT_BODY()
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Basic)
-	TArray<int32> actionArray;
+	TArray<uint8> actionArray;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Basic)
 	FString stateName;
